Create 8d_project.py
The project's purpose is to create an "8D" feeling while listening to songs or audio files.
By creating an artificial impulse response, it can be convolved with the original songs to create a reverb.
By adding to it a division into two stereo channels and making a trigonometric separation between them, a feeling of "movement" between the ears has been created.
A combination of those two mathematical filters into one can lead to an "8D" listening experience.